% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_stems.R
\name{align_stems}
\alias{align_stems}
\title{align two strings}
\usage{
align_stems(right_stem_seq, left_stem_seq, num_loops)
}
\arguments{
\item{right_stem_seq}{string}

\item{left_stem_seq}{a string}

\item{num_loops}{an integer}
}
\value{
a list
}
\description{
align two strings
}
