% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_expand_overhangs}
\alias{calc_expand_overhangs}
\title{calc_expand_overhangs DO NOT USE}
\usage{
calc_expand_overhangs(r1_start, r1_end, r2_start, r2_width)
}
\arguments{
\item{r1_start}{A vector of ints}

\item{r1_end}{A vector of ints}

\item{r2_start}{A vector of ints}

\item{r2_end}{A vector of ints}
}
\value{
A data.frame representing the number of proper and improper overhangs calculated at each shift position
}
\description{
This function takes in vectors of start and stop positions for read groups 1 and 2
It then shifts the positions and checks to see which sets have proper and improper overhangs
Over a larger distance of shifts compared to calc_overhangs
}
