% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_overhangs}
\alias{calc_overhangs}
\title{calc_overhangs}
\usage{
calc_overhangs(
  r1_start,
  r1_end,
  r2_start,
  r2_width,
  dupes_present,
  r1_dupes,
  r2_dupes
)
}
\arguments{
\item{r1_start}{A vector of ints}

\item{r1_end}{A vector of ints}

\item{r2_start}{A vector of ints}

\item{dupes_present}{A bool indicating if vectors of duplicate counts are supplied}

\item{r1_dupes}{A vector of ints representing the number of duplicates}

\item{r2_dupes}{A vector of ints representing the number of duplicates}

\item{r2_end}{A vector of ints}
}
\value{
A data.frame representing the number of proper and improper overhangs calculated at each shift position
}
\description{
This function takes in vectors of start and stop positions for read groups 1 and 2
It then shifts the positions and checks to see which sets have proper and improper overhangs
}
