% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorNtByDepth.R
\name{colorNtByDepth}
\alias{colorNtByDepth}
\title{function to specify colors for each nt as a function of depth}
\usage{
colorNtByDepth(seq_cov_dat)
}
\arguments{
\item{seq_cov_dat}{a data table or data frame}
}
\value{
a data frame
}
\description{
function to specify colors for each nt as a function of depth
}
