% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertU}
\alias{convertU}
\title{convertU}
\usage{
convertU(input_str, size)
}
\arguments{
\item{input_str}{A vector of strings}

\item{size}{An integer}
}
\value{
A vector of strings containing the converted sequences
}
\description{
This function takes in input_str, a vector of strings of nucleotides and converts all 'T's found to 'U's
It also takes in an integer called size to track the size of the number of strings in the vector
It returns the the vector of converted strings
}
