% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicer_overlaps.R
\name{dicer_overlaps}
\alias{dicer_overlaps}
\title{calculates overlaps for Dicer signature}
\usage{
dicer_overlaps(dicer_dt, helix_df, chrom_name, reg_start)
}
\arguments{
\item{dicer_dt}{a data frame}

\item{helix_df}{a data frame}

\item{chrom_name}{a string}

\item{reg_start}{an integer}
}
\value{
all_overlaps
}
\description{
calculates overlaps for Dicer signature
}
