% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased_piRNA.R
\name{.phased_piRNA}
\alias{.phased_piRNA}
\title{phased piRNA function
processes reads according to phased piRNA algorith}
\usage{
.phased_piRNA(
  strand,
  chrom_name,
  reg_start,
  reg_stop,
  input_file,
  logfile,
  wkdir,
  pal,
  plot_output
)
}
\description{
phased piRNA function
processes reads according to phased piRNA algorith
}
