% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps2png.R
\name{eps2png}
\alias{eps2png}
\title{eps2png}
\usage{
eps2png(path_to_magick_exe, file_dir)
}
\arguments{
\item{path_to_magick_exe}{A string specifying the full path to the ImageMagick .exe file}

\item{file_dir}{A string specifying the path to the .eps files to be converted. This will be the location of the converted .png files}
}
\value{
nothing
}
\description{
eps2png
}
