% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists_not_empty.R
\name{exists_not_empty}
\alias{exists_not_empty}
\title{exists_not_empty}
\usage{
exists_not_empty(file_name)
}
\arguments{
\item{file_name}{Filename to be tested. If path is not given, working directory will be the path}
}
\value{
TRUE or FALSE
}
\description{
Checks to see if a given file exists and is not empty
}
\examples{
if (exists_not_empty(file_name)) { print("True") }
}
