% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_df.R
\name{filter_df}
\alias{filter_df}
\title{function to filter dt based on passed in min and max widths
takes two a df and a min and max width
returns a filtered dt}
\usage{
filter_df(df, min_width, max_width)
}
\arguments{
\item{df}{a data frame of read data}

\item{min_width}{a whole number passed in}

\item{max_width}{a whole number passed in}
}
\value{
filtered_df
}
\description{
function to filter dt based on passed in min and max widths
takes two a df and a min and max width
returns a filtered dt
}
