% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dt.R
\name{filter_dt}
\alias{filter_dt}
\title{function to filter dt based on passed in min and max widths
selects only reads beginning with T for piRNAs
takes two a df and a min and max width
returns a filtered dt}
\usage{
filter_dt(chrom)
}
\arguments{
\item{chrom}{a chrom obj}
}
\value{
filtered_df
}
\description{
function to filter dt based on passed in min and max widths
selects only reads beginning with T for piRNAs
takes two a df and a min and max width
returns a filtered dt
}
