% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sh_dt.R
\name{filter_sh_dt}
\alias{filter_sh_dt}
\title{function to filter data table for short hairpin function
makes chrom obj into bam df and filters reads between 18-25 nt
summarizes count of grouped reads
returns filter dt}
\usage{
filter_sh_dt(chrom, chrom_name)
}
\arguments{
\item{chrom}{a chrom object}

\item{chrom_name}{a string}
}
\value{
filter_dt
}
\description{
function to filter data table for short hairpin function
makes chrom obj into bam df and filters reads between 18-25 nt
summarizes count of grouped reads
returns filter dt
}
