% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_overlaps.R
\name{find_overlaps}
\alias{find_overlaps}
\title{function to find regular overlaps
takes two data tables of reads with start and stop positions
converts to gr objects and finds overlaps
returns overlap table}
\usage{
find_overlaps(r1_dt, r2_dt = NULL)
}
\arguments{
\item{r1_dt}{a data frame which contains a read_name, start, and stop}

\item{r2_dt}{a data frame which contains a read_name, start, and stop}
}
\value{
overlap table
}
\description{
function to find regular overlaps
takes two data tables of reads with start and stop positions
converts to gr objects and finds overlaps
returns overlap table
}
