% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_nuc_T.R
\name{first_nuc_T}
\alias{first_nuc_T}
\title{function to calculate the percent of reads that start with T
takes forward and reverse data tables}
\usage{
first_nuc_T(forward_dt, reverse_dt)
}
\arguments{
\item{forward_dt}{a data table of read data}

\item{reverse_dt}{a data table of read data}
}
\value{
num
}
\description{
returns percent value
}
