% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_rna.R
\name{fold_rna}
\alias{fold_rna}
\title{function to run RNAfold 
processes output of RNA fold to get MFE and vien struct
returns list of values for each region}
\usage{
fold_rna(start, stop, converted)
}
\arguments{
\item{start}{a whole number}

\item{stop}{a whole number}

\item{converted}{a vector containing a sequence}
}
\value{
list
}
\description{
function to run RNAfold 
processes output of RNA fold to get MFE and vien struct
returns list of values for each region
}
