% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_short_rna.R
\name{fold_short_rna}
\alias{fold_short_rna}
\title{function to run RNAfold
processes output of RNA fold to get MFE and vien struct
returns list of values for each region}
\usage{
fold_short_rna(start, stop, converted, path_to_RNAfold, chrom_name, wkdir)
}
\arguments{
\item{start}{an integer}

\item{stop}{an integer}

\item{converted}{a vector containing a sequence}

\item{path_to_RNAfold}{a string}

\item{chrom_name}{a string}

\item{wkdir}{a string}
}
\value{
list
}
\description{
function to run RNAfold
processes output of RNA fold to get MFE and vien struct
returns list of values for each region
}
