% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chrom_dna.R
\name{get_chrom_dna}
\alias{get_chrom_dna}
\title{function to subset one chromosome sequence from a genome file
takes genome object
returns vector with dna sequence}
\usage{
get_chrom_dna(genome)
}
\arguments{
\item{genome}{a genome fasta file read by Biostrings}
}
\value{
dna
}
\description{
function to subset one chromosome sequence from a genome file
takes genome object
returns vector with dna sequence
}
