% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GC_content.R
\name{get_GC_content}
\alias{get_GC_content}
\title{function to calculate GC content of reads
takes two data tables of reads with sequences
returns GC percent}
\usage{
get_GC_content(seqs)
}
\arguments{
\item{seqs}{a vector of sequences}
}
\value{
GC percent
}
\description{
function to calculate GC content of reads
takes two data tables of reads with sequences
returns GC percent
}
