% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_pi_heat.R
\name{get_max_pi_heat}
\alias{get_max_pi_heat}
\title{function to find max_pi heatmap cell values
takes a matrix created from piRNA function
finds max row, column, and count of highest cell
returns highest_pi_row, highest_pi_col, second_pi_row, second_pi_col, highest_pi_count, second_pi_count as a data frame}
\usage{
get_max_pi_heat(pi_res)
}
\arguments{
\item{pi_res}{a data table}
}
\value{
max_pi_df
}
\description{
function to find max_pi heatmap cell values
takes a matrix created from piRNA function
finds max row, column, and count of highest cell
returns highest_pi_row, highest_pi_col, second_pi_row, second_pi_col, highest_pi_count, second_pi_count as a data frame
}
