% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_nearby}
\alias{get_nearby}
\title{get_nearby}
\usage{
get_nearby(f_start, f_end, r_start, r_end, max_dist, dfsize)
}
\arguments{
\item{f_start}{A vector of ints}

\item{f_end}{A vector of ints}

\item{r_start}{A vector of ints}

\item{r_end}{A vector of ints}

\item{max_dist}{An integer}

\item{dfsize}{An integer}
}
\value{
A data.frame of nearby reads
}
\description{
This function takes in vectors of start and end positions from 2 sets of reads
It also takes in an integer that sets the maximum distance apart the reads should be and an integer that tracks the size of the input vectors
It returns a data.frame of reads that are nearby - determined by max_dist - but not overlapping
}
