% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuc_10.R
\name{get_nuc_10}
\alias{get_nuc_10}
\title{function to get percentage of reads with A at 10th nucleotide position
takes reverse dt and forward dt
returns percentage}
\usage{
get_nuc_10(forward_dt, reverse_dt)
}
\arguments{
\item{forward_dt}{a data table of read data}

\item{reverse_dt}{a data table of read data}
}
\value{
percentage
}
\description{
function to get percentage of reads with A at 10th nucleotide position
takes reverse dt and forward dt
returns percentage
}
