% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_overlap_counts}
\alias{get_overlap_counts}
\title{get_overlap_counts}
\usage{
get_overlap_counts(
  fdt_start,
  fdt_end,
  fwidth,
  rdt_end,
  rdt_start,
  rwidth,
  check_pi
)
}
\arguments{
\item{fdt_start}{A vector of ints}

\item{fdt_end}{A vector of ints}

\item{fwidth}{An integer}

\item{rdt_end}{A vector of ints}

\item{rdt_start}{A vector of ints}

\item{rwidth}{An integer}

\item{check_pi}{A bool, if TRUE (1), the function will look for piRNA overlaps, if FALSE (0) it will check for siRNA overlaps}
}
\value{
result A matrix representing the overlaps present at each width of reads from 15-32
}
\description{
This function takes in vectors of start and stop positions for read groups 1 and 2
It calculates overlaps for each set of reads and stores them in a matrix based on the read widths
}
