% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pi_overlaps}
\alias{get_pi_overlaps}
\title{get_pi_overlaps}
\usage{
get_pi_overlaps(fdt_start, fdt_end, fwidth, rdt_end, rdt_start, rwidth)
}
\arguments{
\item{fdt_start}{A vector of ints}

\item{fdt_end}{A vector of ints}

\item{fwidth}{An integer}

\item{rdt_end}{A vector of ints}

\item{rdt_start}{A vector of ints}

\item{rwidth}{An integer}
}
\value{
result A matrix representing the overlaps present at each width of reads from 15-32
}
\description{
This function calculates the overlaps present at each width of reads from 15-32
}
