% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_dist.R
\name{get_read_dist}
\alias{get_read_dist}
\title{function to calculate the read distribution
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution}
\usage{
get_read_dist(forward_dt, reverse_dt)
}
\arguments{
\item{forward_dt}{a data table containing read data}

\item{reverse_dt}{a data table containing read data}
}
\value{
dizeDist
}
\description{
function to calculate the read distribution
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution
}
