% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_pileups.R
\name{get_read_pileups}
\alias{get_read_pileups}
\title{A function to extract read pileups at specified start and stop}
\usage{
get_read_pileups(start, stop, bam_scan, input_file)
}
\arguments{
\item{start}{a whole number}

\item{stop}{a whole number}

\item{bam_scan}{a bam scan object}

\item{input_file}{a string}
}
\value{
a data table
}
\description{
A function to extract read pileups at specified start and stop
}
