% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_size_dist.R
\name{get_read_size_dist}
\alias{get_read_size_dist}
\title{get_read_size_dist
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution}
\usage{
get_read_size_dist(f_df, r_df)
}
\arguments{
\item{f_df}{a data.frame of forward reads}

\item{r_df}{a data.frame of reverse reads}
}
\value{
size_dist
}
\description{
get_read_size_dist
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution
}
