% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_string.R
\name{get_region_string}
\alias{get_region_string}
\title{Function to revert the bed file positions back to zero based for writing results}
\usage{
get_region_string(chrom, start, stop)
}
\arguments{
\item{chrom}{Character: The region name}

\item{start}{Integer: The one based region start position}

\item{stop}{Integer: The one based region stop position}
}
\value{
Character: The region name and zero based positions
}
\description{
Function to revert the bed file positions back to zero based for writing results
}
\examples{
get_region_string("Dsim_2L", 1, 75541)

}
