% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stranded_read_dist.R
\name{get_stranded_read_dist}
\alias{get_stranded_read_dist}
\title{get_stranded_read_dist
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution}
\usage{
get_stranded_read_dist(bam_obj, chrom_name, reg_start, reg_stop)
}
\arguments{
\item{bam_obj}{an object created by Rsamtools}

\item{chrom_name}{a string}

\item{reg_start}{an integer}

\item{reg_stop}{an integer}
}
\value{
size_dist
}
\description{
get_stranded_read_dist
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution
}
