% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_n.R
\name{get_top_n}
\alias{get_top_n}
\title{function to filter data table for short hairpin function
makes chrom obj into bam df and filters reads between 18-25 nt
summarizes count of grouped reads
returns top 10% of reads}
\usage{
get_top_n(chrom, chrom_name, n)
}
\arguments{
\item{chrom}{a chrom object}

\item{chrom_name}{a string}

\item{n}{a decimal}
}
\value{
filter_dt
}
\description{
function to filter data table for short hairpin function
makes chrom obj into bam df and filters reads between 18-25 nt
summarizes count of grouped reads
returns top 10% of reads
}
