% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_n_weighted.R
\name{get_top_n_weighted}
\alias{get_top_n_weighted}
\title{function takes a data table of reads
summarizes count of grouped reads
returns a weighted top n\% of reads}
\usage{
get_top_n_weighted(dt, chrom_name, n)
}
\arguments{
\item{dt}{a data table of reads}

\item{chrom_name}{a string}

\item{n}{a decimal representing percent of reads to return}
}
\value{
filter_dt
}
\description{
function takes a data table of reads
summarizes count of grouped reads
returns a weighted top n\% of reads
}
