% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weighted_read_dist.R
\name{get_weighted_read_dist}
\alias{get_weighted_read_dist}
\title{function to calculate the read distribution
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution}
\usage{
get_weighted_read_dist(forward_dt, reverse_dt)
}
\arguments{
\item{forward_dt}{a minus strand object}

\item{reverse_dt}{a plus strand object}
}
\value{
dizeDist
}
\description{
function to calculate the read distribution
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution
}
