% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChrP_allbed.R
\name{getChrP_allbed}
\alias{getChrP_allbed}
\title{function to extract all reads in specified regions on minus strand
takes bam object, chrom_list, start_list, and stop_list
extracts rname, pos, qwidth, from bam obj}
\usage{
getChrP_allbed(bam_obj, chrom_list, start_list, stop_list)
}
\arguments{
\item{bam_obj}{a bam obj created by openBamFile}

\item{chrom_list}{a list of chromosome names from column 1 of bed file}

\item{start_list}{a list of whole numbers from column 2 of bedfile}

\item{stop_list}{a list of whole numbers from column 3 of bed file}
}
\value{
chrom obj
}
\description{
function to extract all reads in specified regions on minus strand
takes bam object, chrom_list, start_list, and stop_list
extracts rname, pos, qwidth, from bam obj
}
