% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChrPlus.R
\name{getChrPlus}
\alias{getChrPlus}
\title{create an object from read data from plus strand
takes bam obj, chrom name, reg_start, and reg_stop}
\usage{
getChrPlus(bam_obj, chrom_name, reg_start, reg_stop)
}
\arguments{
\item{bam_obj}{a bam object created by openBamFile}

\item{reg_start}{an integer}

\item{reg_stop}{an integer}
}
\value{
max_overhang
}
\description{
create an object from read data from plus strand
takes bam obj, chrom name, reg_start, and reg_stop
}
