% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLoopPileupsCPP}
\alias{getLoopPileupsCPP}
\title{getLoopPileupsCPP}
\usage{
getLoopPileupsCPP(
  r1Start,
  r1Stop,
  lStart,
  lStop,
  r2Start,
  r2Stop,
  dt_col_1,
  dt_col_2,
  total_count
)
}
\arguments{
\item{r1Start}{A vector of ints}

\item{r1Stop}{A vector of ints}

\item{lStart}{A vector of ints}

\item{lStop}{A vector of ints}

\item{r2Start}{A vector of ints}

\item{r2Stop}{A vector of ints}

\item{dt_col_1}{A vector of ints}

\item{dt_col_2}{A vector of ints}

\item{int}{total_count An integer}
}
\value{
df A data.frame
}
\description{
This function returns a data.frame containing the start and stop positions of read 1, read 2, and a loop
If the number of pileups in a loop divided by the total number of pileups is less that 0.05
}
