% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPileups}
\alias{getPileups}
\title{getPileups}
\usage{
getPileups(dtpos, dtcount, start_r1, end_r1, start_r2, end_r2)
}
\arguments{
\item{input_str}{A vector of ints}

\item{start}{A vector of ints}

\item{stop}{A vector of ints}

\item{size}{An integer}
}
\value{
A data.frame consisting of start and stop positions along with their fasta sequence
}
\description{
This function takes in a fasta sequence as a string, an int vector of start positions, an int vector of stop positions,
and an int to track the size of the input vectors.
It returns a data.frame containing each input start and stop positions along with the fasta sequence associated with those positions
}
