% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPileupsMap}
\alias{getPileupsMap}
\title{getPileupsMap}
\usage{
getPileupsMap(dtpos, dtcount, start_r1, end_r1, start_r2, end_r2)
}
\arguments{
\item{dtpos}{A vector of ints representing positions}

\item{dtcount}{A vector of ints representing the pileups at each dtpos position}

\item{start_r1}{A vector of start positions from read group 1}

\item{end_r1}{A vector of end positions from read group 2}

\item{start_r2}{A vector of start positions from read group 2}

\item{end_r2}{A vector of end positions from read group 2}

\item{count}{A vector of ints representing the number of duplicates each start_r1, end_r1, start_r2, end_r2 had prior to grouping}
}
\value{
A data.frame consisting of the average pileups in read 1 range and in read 2 range
}
\description{
This function takes in 2 vectors of calculated pileups and their positions.
It also takes in vectors representing start and end positions of read group 1 and 2 with a vector tracking their duplicates
It converts the pileups vectors into a std::map and calculates the average pileups in a particular read range and returns it in a data.frame
}
