% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{group_helix_res}
\alias{group_helix_res}
\title{group_helix_res}
\usage{
group_helix_res(x, y)
}
\arguments{
\item{x}{A vector of ints}

\item{y}{A vector of ints}
}
\value{
A data.frame consisting of start and stop positions for both x and y
}
\description{
This function takes in two vectors of paired positions. It iterates through each and tests to see if a the
positions have changed by more than 4 nt in order to track potential hairpin loops.
It returns a data.frame consisting of x start and stop positions along with their respective y start and stop positions
}
