% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hairpin.R
\name{hairpin}
\alias{hairpin}
\title{run the hairpin function
runs hairpin function on both strands
takes a gtf file, bam file, and genome file specified by user
takes min_read_count specified by user. Default = 1}
\usage{
hairpin(vars)
}
\arguments{
\item{vars}{a list}
}
\value{
plots
}
\description{
run the hairpin function
runs hairpin function on both strands
takes a gtf file, bam file, and genome file specified by user
takes min_read_count specified by user. Default = 1
}
