% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hairpin_function.R
\name{hairpin_function}
\alias{hairpin_function}
\title{function to run the hairpin algorithm
processes reads from bam object according to strand
plots the arc plot and read dist}
\usage{
hairpin_function(
  chrom_name,
  reg_start,
  reg_stop,
  bam_obj,
  bam_header,
  chr_name,
  chr_length,
  chromosome,
  strand,
  min_read_count,
  genome_file,
  input_file
)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{bam_obj}{a bam object created by openBamFile}

\item{bam_header}{header information from bam file}

\item{chr_name}{chrom names extracted from bam file}

\item{chr_length}{length of chromosome of interest}

\item{chromosome}{index of chromosome of interest}

\item{strand}{a character passed in, "+" or "-"}

\item{min_read_count}{a whole number}

\item{genome_file}{a fasta file of chrom sequences}

\item{input_file}{a BAM file}
}
\value{
max_overhang
}
\description{
function to run the hairpin algorithm
processes reads from bam object according to strand
plots the arc plot and read dist
}
