% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_sizes.R
\name{highest_sizes}
\alias{highest_sizes}
\title{function to find the read size of hte highest abundance
takes a table of read sizes and counts
returns highest and second highest read sizes}
\usage{
highest_sizes(read_dist)
}
\arguments{
\item{read_dist}{a table}
}
\value{
a list
}
\description{
function to find the read size of hte highest abundance
takes a table of read sizes and counts
returns highest and second highest read sizes
}
