% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_vector.R
\name{insert_vector}
\alias{insert_vector}
\title{insert vector
inserts characters into string vector at desired position}
\usage{
insert_vector(input, position, values)
}
\arguments{
\item{input}{a string vector}

\item{position}{an integer}

\item{values}{a character}
}
\value{
res a vector
}
\description{
insert vector
inserts characters into string vector at desired position
}
