% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_count_table}
\alias{make_count_table}
\title{make_count_table}
\usage{
make_count_table(fdt_start, fdt_end, fwidth, rdt_start, rdt_end, rwidth)
}
\arguments{
\item{fdt_start}{A vector of ints}

\item{fdt_end}{A vector of ints}

\item{fwidth}{An integer}

\item{rdt_start}{A vector of ints}

\item{rdt_end}{A vector of ints}

\item{rwidth}{An integer}
}
\value{
df A data.frame containing the counts of each overlap to be used for a z score
}
\description{
This functions calculates the counts of each overlap to be used for z score
}
