% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_helix.R
\name{make_helix}
\alias{make_helix}
\title{function to convert vienna to helix}
\usage{
make_helix(fold_list)
}
\arguments{
\item{fold_list}{a list}
}
\value{
list
}
\description{
function to convert vienna to helix
}
