% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_si_BamDF.R
\name{make_si_BamDF}
\alias{make_si_BamDF}
\title{function to convert chrom obj into a data frame
extracts rname, pos, width, and seq for siRNA dataframe}
\usage{
make_si_BamDF(chrom_obj)
}
\arguments{
\item{chrom_obj}{an object created from a bam obj}
}
\value{
bam_df
}
\description{
function to convert chrom obj into a data frame
extracts rname, pos, width, and seq for siRNA dataframe
}
