% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBamDF.R
\name{makeBamDF}
\alias{makeBamDF}
\title{function to convert chrom obj into a data frame
extracts rname, pos, width, seq, and the first letter of the sequence
finds max row, column, and count of highest cell}
\usage{
makeBamDF(chrom_obj)
}
\arguments{
\item{chrom_obj}{an object created from a bam obj}
}
\value{
bam_df
}
\description{
function to convert chrom obj into a data frame
extracts rname, pos, width, seq, and the first letter of the sequence
finds max row, column, and count of highest cell
}
