% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRobj.R
\name{makeGRobj}
\alias{makeGRobj}
\title{function to create a GR obj from a data table of reads
needed for finding overlaps
returns a GR obj}
\usage{
makeGRobj(data, name, reg_start, reg_stop)
}
\arguments{
\item{data}{a data table}

\item{name}{a string passed in by user or bed file}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}
}
\value{
GRobj
}
\description{
function to create a GR obj from a data table of reads
needed for finding overlaps
returns a GR obj
}
