% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{map_and_count}
\alias{map_and_count}
\title{map_and_count}
\usage{
map_and_count(fstart, fstart_size, rend, proper_overlap)
}
\arguments{
\item{fstart}{A vector of ints}

\item{fstart_size}{An integer}

\item{rend}{A vector of ints}

\item{proper_overlap}{An integer}
}
\value{
An integer represent the number of overlaps present between the input vectors
}
\description{
This function takes in a vector of start positions from read group 1 and
a vector of stop positions from read group 2 along with an overlap width
It calculates counts proper overlaps contained in the vectors and returns the count
}
