% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mergePileups}
\alias{mergePileups}
\title{mergePileups}
\usage{
mergePileups(start, end, pileups_pos, pileups_count)
}
\arguments{
\item{start}{A vector of ints}

\item{end}{A vector of ints}

\item{pileups_pos}{A vector of ints}

\item{pileups_count}{A vector of ints}
}
\value{
A data.frame consisting of positions and the counts associated with those positions
}
\description{
This function takes in vectors of start and end positions and pileups positions and their counts
It returns a dataframe with only the positions from start and the related counts as found in pileups_count
}
