% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA.R
\name{miRNA}
\alias{miRNA}
\title{miRNA function
calls the .miRNA function for both strands, creates miRNA logfile}
\usage{
miRNA(vars)
}
\arguments{
\item{vars}{a list of variables provided by user}
}
\value{
plots
}
\description{
miRNA function
calls the .miRNA function for both strands, creates miRNA logfile
}
