% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA_function.R
\name{miRNA_function}
\alias{miRNA_function}
\title{run miRNA function
calls the .miRNA function for both strands, creates miRNA logfile}
\usage{
miRNA_function(vars)
}
\arguments{
\item{vars}{a list of variables provided by user}
}
\value{
plots
}
\description{
run miRNA function
calls the .miRNA function for both strands, creates miRNA logfile
}
