% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misi_pi.R
\name{misi_pi}
\alias{misi_pi}
\title{runs piRNA_function and creates dir and logfile for outputs}
\usage{
misi_pi(vars)
}
\arguments{
\item{vars}{a list}
}
\value{
pi_res, plots
}
\description{
runs piRNA_function and creates dir and logfile for outputs
}
