% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSiPi.RNA-package.R
\docType{package}
\name{MiSiPi.RNA}
\alias{MiSiPi.RNA}
\alias{MiSiPi.RNA-package}
\title{MiSiPi.RNA: Characterization of Small RNA Pathways}
\description{
Creates plots and analyzes small RNA sequencing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stupornova33/MiSiPi.RNA}
}

}
\author{
\strong{Maintainer}: Taiya Jarva \email{t.m.jarva13@gmail.com}

Other contributors:
\itemize{
  \item Chris Baugh \email{chris.baugh@gmail.com} [contributor]
}

}
\keyword{internal}
