% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misipi_rna.R
\name{misipi_rna}
\alias{misipi_rna}
\title{Wrapper function that calls core functions}
\usage{
misipi_rna(vars, method = c("all", "miRNA", "piRNA", "siRNA"))
}
\arguments{
\item{vars}{a list of parameters created by set_vars}

\item{method}{a string indicating which processing method should be used}
}
\value{
plots
}
\description{
Wrapper function that calls core functions
}
