% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_get_si_overlaps}
\alias{new_get_si_overlaps}
\title{new_get_si_overlaps}
\usage{
new_get_si_overlaps(fdt_start, fdt_end, fwidth, rdt_start, rdt_end, rwidth)
}
\arguments{
\item{fdt_start}{A vector of ints}

\item{fdt_end}{A vector of ints}

\item{fwidth}{An integer}

\item{rdt_start}{A vector of ints}

\item{rdt_end}{A vector of ints}

\item{rwidth}{An integer}
}
\value{
result A matrix representing the overlaps present at each width of reads from 15-32
}
\description{
This function takes in vectors of start and stop positions for read groups 1 and 2
It calculates overlaps for each set of reads and stores them in a matrix based on the read widths
}
