% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_miRNA_function.R
\name{new_miRNA_function}
\alias{new_miRNA_function}
\title{new_miRNA_function
processes reads according to miRNA_algorithm, new with rnaplot function
returns plots}
\usage{
new_miRNA_function(
  chrom_name,
  reg_start,
  reg_stop,
  chromosome,
  length,
  strand,
  min_read_count,
  genome_file,
  bam_file,
  logfile,
  wkdir,
  plot_output,
  path_to_RNAfold,
  path_to_RNAplot,
  write_fastas,
  weight_reads,
  out_type
)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{chromosome}{chrom names extracted from bam file}

\item{length}{length of chromosome of interest}

\item{strand}{a character passed in, "+" or "-"}

\item{min_read_count}{a whole number}

\item{genome_file}{a fasta file of chrom sequences}

\item{bam_file}{a BAM file}

\item{logfile}{a string}

\item{wkdir}{a string}

\item{plot_output}{a bool, default = TRUE}

\item{path_to_RNAfold}{a string}

\item{path_to_RNAplot}{a string}

\item{write_fastas}{a bool, TRUE or FALSE}

\item{weight_reads}{Determines whether read counts will be weighted and with which method. Valid options are "weight_by_prop", "locus_norm", a user-defined value, or "none". See MiSiPi documentation for descriptions of the weighting methods.}

\item{out_type}{The type of file to write the plots to. Options are "png" or "pdf". Default is PDF.}
}
\value{
plots
}
\description{
new_miRNA_function
processes reads according to miRNA_algorithm, new with rnaplot function
returns plots
}
