% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_run_all.R
\name{new_run_all}
\alias{new_run_all}
\title{run_all function
run all modules of the package and store metrics for ML}
\usage{
new_run_all(
  chrom_name,
  reg_start,
  reg_stop,
  chromosome,
  length,
  bam_file,
  roi,
  genome_file,
  min_read_count,
  si_pal,
  pi_pal,
  plot_output,
  path_to_RNAfold,
  annotate_bed,
  weight_reads,
  bed_file
)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{an integer}

\item{reg_stop}{an integer}

\item{chromosome}{an integer representing the chromosome number}

\item{length}{an integer}

\item{bam_file}{a string}

\item{roi}{a string}

\item{genome_file}{a string}

\item{min_read_count}{an integer}

\item{si_pal}{a string}

\item{pi_pal}{a string}

\item{plot_output}{a string, "T" or "F"}

\item{path_to_RNAfold}{a string}

\item{annotate_bed}{a string, "T" or "F"}

\item{weight_reads}{a string, "T", or "F"}

\item{bed_file}{a string}
}
\value{
results
}
\description{
run_all function
run all modules of the package and store metrics for ML
}
