% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_hp_res.R
\name{null_hp_res}
\alias{null_hp_res}
\title{This function creates default null results to return in the event that no meaningful results could be created}
\usage{
null_hp_res()
}
\value{
list
}
\description{
This function creates default null results to return in the event that no meaningful results could be created
}
