% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_mi_res.R
\name{null_mi_res}
\alias{null_mi_res}
\title{null_mi_res}
\usage{
null_mi_res()
}
\value{
default null result object
}
\description{
creates a default object for return to miRNA_function in the event that no meaningful results could be calculated in a particular iteration
}
