% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_loop.R
\name{one_loop}
\alias{one_loop}
\title{run in process miRNA if one loop}
\usage{
one_loop(vienna_vec, orig_loop_pos, converted)
}
\arguments{
\item{vienna_vec}{a string}

\item{orig_loop_pos}{a string}

\item{converted}{a string}
}
\value{
res_list a list
}
\description{
run in process miRNA if one loop
}
