% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_readsize_dist.R
\name{output_readsize_dist}
\alias{output_readsize_dist}
\title{Make read size distribution table}
\usage{
output_readsize_dist(dat, prefix, wkdir, strand, type)
}
\arguments{
\item{dat}{a data frame containing read sizes and counts}

\item{prefix}{a string}

\item{wkdir}{a string}

\item{strand}{a string, "+" or "-". Default is NULL.}

\item{type}{a string}
}
\value{
nothing
}
\description{
Make read size distribution table
}
