% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhang_z_plot.R
\name{overhang_z_plot}
\alias{overhang_z_plot}
\title{plots the z_score for the 2nt overhang
takes a table of overhang z_scores
only output is overhang plots}
\usage{
overhang_z_plot(overhangs)
}
\arguments{
\item{overhangs}{a table}
}
\value{
plot
}
\description{
plots the z_score for the 2nt overhang
takes a table of overhang z_scores
only output is overhang plots
}
