% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overlap_counts}
\alias{overlap_counts}
\title{overlap_counts}
\usage{
overlap_counts(f_start, X_SIZE, r_end, Y_SIZE, overlap)
}
\arguments{
\item{f_start}{A vector of ints}

\item{X_SIZE}{An Integer}

\item{r_end}{A vector of ints}

\item{Y_SIZE}{An Integer}

\item{overlap}{An Integer}
}
\value{
count An Integer representing the number of overlaps present
}
\description{
This function takes in vectors of start positions from read group 1 and end positions from read group 2
It also takes in an overlap amount integer
The function counts the number of overlaps occurring and returns them as an integer
}
