% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased_piRNA.R
\name{phased_piRNA}
\alias{phased_piRNA}
\title{phased_piRNA
calls phased_piRNA function on both strand, creates output wkdir and logfile}
\usage{
phased_piRNA(vars)
}
\arguments{
\item{vars}{a list}
}
\value{
plots
}
\description{
phased_piRNA
calls phased_piRNA function on both strand, creates output wkdir and logfile
}
