% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased_piRNA_function.R
\name{phased_piRNA_function}
\alias{phased_piRNA_function}
\title{phased piRNA function
processes reads according to phased piRNA algorith
plots output
takes bam object and strand 
returns max_zscore and plots}
\usage{
phased_piRNA_function(bam_obj, strand)
}
\arguments{
\item{bam_obj}{an object created by openBamFile}

\item{strand}{a character passed in, either "+" or "-"}
}
\value{
max_zscore
}
\description{
phased piRNA function
processes reads according to phased piRNA algorith
plots output
takes bam object and strand 
returns max_zscore and plots
}
