% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piRNA.R
\name{piRNA}
\alias{piRNA}
\title{runs .piRNA and creates dir and logfile for outputs}
\usage{
piRNA(vars)
}
\arguments{
\item{vars}{a list}
}
\value{
pi_res, plots
}
\description{
runs .piRNA and creates dir and logfile for outputs
}
