% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piRNA_function.R
\name{piRNA_function}
\alias{piRNA_function}
\title{run the piRNA function
processes forward and reverse reads according to piRNA algorithm
outputs plots, heat results, and zdf}
\usage{
piRNA_function(forward_dt, reverse_dt)
}
\arguments{
\item{forward_dt}{a data table containing read data}

\item{reverse_dt}{a data table containing read data}
}
\value{
plots, heat results, and zdf
}
\description{
run the piRNA function
processes forward and reverse reads according to piRNA algorithm
outputs plots, heat results, and zdf
}
