% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bed.R
\name{plot_bed}
\alias{plot_bed}
\title{takes a gff file
outputs a plot}
\usage{
plot_bed(gff_path, chrom_name, reg_start, reg_stop)
}
\arguments{
\item{gff_path}{a string path leading to the gff file}

\item{chrom_name}{a string}

\item{reg_start}{an integer}

\item{reg_stop}{an integer}
}
\value{
plot
}
\description{
takes a gff file
outputs a plot
}
