% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_coverage}
\alias{plot_coverage}
\title{plots the coverage over an interval}
\usage{
plot_coverage(dt)
}
\arguments{
\item{dt}{a dataframe}
}
\value{
a histogram plot
}
\description{
plots the coverage over an interval
}
