% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{function to plot densities of different read sizes}
\usage{
plot_density(all_df, reg_start, reg_stop)
}
\arguments{
\item{all_df}{a data frame}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}
}
\value{
density
}
\description{
function to plot densities of different read sizes
}
