% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gtf.R
\name{plot_gtf}
\alias{plot_gtf}
\title{takes a gtf file
outputs a plot}
\usage{
plot_gtf(gtf_file, chrom_name, reg_start, reg_stop)
}
\arguments{
\item{gtf_file}{a string path leading to the gtf file}

\item{chrom_name}{a string}

\item{reg_start}{an integer}

\item{reg_stop}{an integer}
}
\value{
plot
}
\description{
takes a gtf file
outputs a plot
}
