% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_large_density.R
\name{plot_large_density}
\alias{plot_large_density}
\title{Plots read sizes over large loci
Faster than geom_area but doesn't look as good}
\usage{
plot_large_density(data, reg_start, reg_stop)
}
\arguments{
\item{data}{a dataframe consisting of three columns, "Position", "Count" and "Length".}

\item{reg_start}{The start position of the region of interest.}

\item{reg_stop}{The end position of a region of interest.}
}
\value{
Density Plot
}
\description{
Plots read sizes over large loci
Faster than geom_area but doesn't look as good
}
