% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_miRNA_cov.R
\name{plot_miRNA_cov}
\alias{plot_miRNA_cov}
\title{plot miRNA coverage
takes a data frame of coverage over a sequence of nucleotides}
\usage{
plot_miRNA_cov(cov_dat, prefix)
}
\arguments{
\item{cov_dat}{a table containing coverage over a nucleotide sequence}

\item{prefix}{a name for the output file}
}
\value{
plots
}
\description{
plot miRNA coverage
takes a data frame of coverage over a sequence of nucleotides
}
