% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_miRNA_struct.R
\name{plot_miRNA_struct}
\alias{plot_miRNA_struct}
\title{plot_miRNA structure}
\usage{
plot_miRNA_struct(f_df)
}
\arguments{
\item{f_df}{a data frame}
}
\value{
plot
}
\description{
plot_miRNA structure
}
