% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overhangz.R
\name{plot_overhangz}
\alias{plot_overhangz}
\title{function to plot the zscore for Dicer overhangs}
\usage{
plot_overhangz(overhang1, overhang2 = NULL)
}
\arguments{
\item{overhang1}{a dataframe}

\item{overhang2}{a dataframe}
}
\value{
plot
}
\description{
function to plot the zscore for Dicer overhangs
}
