% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlapz.R
\name{plot_overlapz}
\alias{plot_overlapz}
\title{plot the overlap probability}
\usage{
plot_overlapz(z_df)
}
\arguments{
\item{z_df}{a data frame}
}
\value{
a ggplot object
}
\description{
plot the overlap probability
}
