% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_si_heat.R
\name{plot_si_heat}
\alias{plot_si_heat}
\title{plot the si heatmaps
takes a matrix as input and a specified color palette
outputs plots}
\usage{
plot_si_heat(
  results,
  pal = c("RdYlBl", "yelOrRed", "MagYel", "Greens", "BlYel")
)
}
\arguments{
\item{bed_file}{a tab-delineated file containing three columns: chrom name, start, and stop}

\item{input_file}{a BAM file}

\item{genome}{a fasta file containing chromosome sequences}

\item{min_read_count}{a whole number}
}
\value{
plots
}
\description{
plot the si heatmaps
takes a matrix as input and a specified color palette
outputs plots
}
