% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sizes.R
\name{plot_sizes}
\alias{plot_sizes}
\title{plots the size distribution of reads}
\usage{
plot_sizes(dist, chrom_name, reg_start, reg_stop)
}
\arguments{
\item{dist}{a data table}

\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}
}
\value{
plots
}
\description{
plots the size distribution of reads
}
