% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sizes_by_strand.R
\name{plot_sizes_by_strand}
\alias{plot_sizes_by_strand}
\title{plots the size distribution of reads separately by strand}
\usage{
plot_sizes_by_strand(chrom_name, reg_start, reg_stop, bam_file, libsize)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{a numerical value}

\item{reg_stop}{a numerical value}

\item{bam_file}{a string}

\item{libsize}{a numeric value}
}
\value{
A plot
}
\description{
plots the size distribution of reads separately by strand
}
