% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text.R
\name{plot_text}
\alias{plot_text}
\title{plots the dot bracket structure above the nucleotide sequence}
\usage{
plot_text(vienna, sequence)
}
\arguments{
\item{vienna}{a string}

\item{sequence}{a string}
}
\value{
a plot
}
\description{
plots the dot bracket structure above the nucleotide sequence
}
