% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_miRNA_struct.R
\name{process_miRNA_struct}
\alias{process_miRNA_struct}
\title{miRNA structure processing function}
\usage{
process_miRNA_struct(vienna, converted)
}
\arguments{
\item{vienna}{a string}

\item{converted}{a string}
}
\value{
f_df
}
\description{
miRNA structure processing function
}
