% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proper_overlap}
\alias{proper_overlap}
\title{proper_overlap}
\usage{
proper_overlap(widthx, widthy)
}
\arguments{
\item{widthx}{An Integer}

\item{widthy}{An Integer}
}
\value{
proper_overlap An integer representing what the proper overlap should be for these positions
}
\description{
This function takes in two integers representing widths and returns a proper overlap
}
