% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_densityBySize.R
\name{read_densityBySize}
\alias{read_densityBySize}
\title{function to filter reads by size and plot pileup density}
\usage{
read_densityBySize(
  bam_obj,
  chrom_name,
  reg_start,
  reg_stop,
  input_file,
  wkdir,
  logfile
)
}
\arguments{
\item{bam_obj}{a string}

\item{chrom_name}{a string}

\item{reg_start}{a string}

\item{reg_stop}{a string}

\item{input_file}{a string}

\item{wkdir}{a string}

\item{logfile}{a string}
}
\value{
all_df
}
\description{
function to filter reads by size and plot pileup density
}
