% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_reads.R
\name{rep_reads}
\alias{rep_reads}
\title{replicates reads from counts dt}
\usage{
rep_reads(counts_dt)
}
\arguments{
\item{counts_dt}{a data table or data frame}
}
\value{
res_df
}
\description{
replicates reads from counts dt
}
