% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rep_seq_reads}
\alias{rep_seq_reads}
\title{rep_seq_reads}
\usage{
rep_seq_reads(rep_count, rname, start, end, first, seq)
}
\arguments{
\item{rep_count}{a vector of ints}

\item{rname}{a vector of strings}

\item{start}{a vector of ints}

\item{end}{a vector of ints}

\item{first}{a vector of characters}

\item{seq}{a vector of strings}
}
\value{
df a dataframe
}
\description{
Takes a vector of "collapsed" reads with a count value and replicates each read as many times as it occurs
}
