% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_plot.R
\name{rna_plot}
\alias{rna_plot}
\title{rna_plot
calls rnafold and rnaplot from viennaRNA and colors the most abundant reads on the structure.}
\usage{
rna_plot(
  path_to_RNAfold,
  path_to_RNAplot,
  wkdir,
  pos_df,
  colors,
  chrom_name,
  reg_start,
  reg_stop,
  most_abundant_start,
  most_abundant_stop,
  strand
)
}
\arguments{
\item{path_to_RNAfold}{a string, the full path to the RNAfold executable.}

\item{path_to_RNAplot}{a string, the full path to the RNAplot executable.}

\item{wkdir}{a string, the path to the desired working directory.}

\item{pos_df}{a data.frame containing the starts and ends of the abundant reads.}

\item{colors}{a list of two colors for highlighting abundant reads on the structure.}

\item{chrom_name}{a string}

\item{reg_start}{an integer}

\item{reg_stop}{an integer}

\item{most_abundant_start}{an integer}

\item{most_abundant_stop}{an integer}

\item{strand}{a character}
}
\value{
nothing
}
\description{
rna_plot
calls rnafold and rnaplot from viennaRNA and colors the most abundant reads on the structure.
}
