% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hairpin_function.R
\name{run_hairpin_function}
\alias{run_hairpin_function}
\title{run the hairpin function
runs hairpin function on both strands
takes a bed file, bam file, and genome file specified by user
takes min_read_count specified by user. Default = 1}
\usage{
run_hairpin_function(vars)
}
\arguments{
\item{vars}{a list}
}
\value{
plots
}
\description{
run the hairpin function
runs hairpin function on both strands
takes a bed file, bam file, and genome file specified by user
takes min_read_count specified by user. Default = 1
}
