% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_miRNA_function.R
\name{run_miRNA_function}
\alias{run_miRNA_function}
\title{plot miRNA coverage
takes a data frame of coverage over a sequence of nucleotides}
\usage{
run_miRNA_function(bed_file, input_file, genome_file)
}
\arguments{
\item{bed_file}{a tab-delineated file with three columns: chrom, start, stop}

\item{input_file}{a BAM file}

\item{genome_file}{a fasta file with chromosome sequences}
}
\value{
plots
}
\description{
plot miRNA coverage
takes a data frame of coverage over a sequence of nucleotides
}
