% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_phased_piRNA_function.R
\name{run_phased_piRNA}
\alias{run_phased_piRNA}
\title{run_phased_piRNA
calls phased_piRNA function on both strand, creates output wkdir and logfile}
\usage{
run_phased_piRNA(vars)
}
\arguments{
\item{vars}{a list}
}
\value{
plots
}
\description{
run_phased_piRNA
calls phased_piRNA function on both strand, creates output wkdir and logfile
}
