% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_piRNA_function.R
\name{run_piRNA_function}
\alias{run_piRNA_function}
\title{run_piRNA_function
takes test_list and input_file
returns pi_res and plots}
\usage{
run_piRNA_function(bed_file, input_file)
}
\arguments{
\item{bed_file}{a tab-delineated file with three columns. Chrom, start, stop}

\item{input_file}{a BAM file}
}
\value{
pi_res, plots
}
\description{
run_piRNA_function
takes test_list and input_file
returns pi_res and plots
}
