% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_siRNA_function.R
\name{run_siRNA_function}
\alias{run_siRNA_function}
\title{run_siRNA_function}
\usage{
run_siRNA_function(bed_file, input_file)
}
\arguments{
\item{bed_file}{a tab-delineated file with three columns: chrom start stop}

\item{input_file}{a BAM file}
}
\value{
si_res
}
\description{
run_siRNA_function
}
