% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vars.R
\name{set_vars}
\alias{set_vars}
\title{takes a gff file
only output is hairpin plots}
\usage{
set_vars(
  roi,
  bam_file,
  genome,
  min_read_count,
  plot_output,
  path_to_RNAfold,
  pi_pal,
  si_pal,
  annotate_bed,
  weight_reads,
  bed_file = NULL
)
}
\arguments{
\item{roi}{a string}

\item{bam_file}{a string}

\item{genome}{a string}

\item{min_read_count}{an integer}

\item{plot_output}{a string, "T" or "F"}

\item{path_to_RNAfold}{a string}

\item{pi_pal}{a string}

\item{si_pal}{a string}

\item{annotate_bed}{a string, "T" or "F"}

\item{weight_reads}{a string, "T" or "F"}

\item{bed_file}{a string}
}
\value{
a list
}
\description{
takes a gff file
only output is hairpin plots
}
