% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siRNA.R
\name{siRNA}
\alias{siRNA}
\title{run .siRNA function
calls the siRNA function, creates siRNA logfile}
\usage{
siRNA(vars)
}
\arguments{
\item{vars}{a list of variables provided by user}
}
\value{
si_res
}
\description{
run .siRNA function
calls the siRNA function, creates siRNA logfile
}
