% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siRNA_function.R
\name{siRNA_function}
\alias{siRNA_function}
\title{siRNA function
process forward and reverse dt and plot si heatmap
takes forward and reverse dt}
\usage{
siRNA_function(forward_dt, reverse_dt)
}
\arguments{
\item{forward_dt}{a data table containing read data}

\item{reverse_dt}{a data table containing read data}
}
\value{
results
}
\description{
siRNA function
process forward and reverse dt and plot si heatmap
takes forward and reverse dt
}
