% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_loops.R
\name{three_loops}
\alias{three_loops}
\title{run in process miRNA if three loops}
\usage{
three_loops(vienna_vec, orig_loop_pos, converted)
}
\arguments{
\item{vienna_vec}{a string}

\item{orig_loop_pos}{a string}

\item{converted}{a string}
}
\value{
res_list a list
}
\description{
run in process miRNA if three loops
}
