% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorsToMap}
\alias{vectorsToMap}
\title{vectorsToMap}
\usage{
vectorsToMap(k, v)
}
\arguments{
\item{k}{A referenced vector of integers representing keys}

\item{v}{A referenced vector of integers representing values}
}
\value{
A std::map of int pairs
}
\description{
This function takes takes in references to vectors of keys and values
These keys and values are turned into pairs in a std::map and returned
This function assumes the vectors are of the same size.
This is only intended to be called from getPileupsMap
}
