% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_by_prop.R
\name{weight_by_prop}
\alias{weight_by_prop}
\title{function takes a data table of reads
summarizes count of grouped reads
returns dataframe of top reads replicated to reflect their proportional weight to the top read}
\usage{
weight_by_prop(dt, chrom_name)
}
\arguments{
\item{dt}{a data table of reads}

\item{chrom_name}{a string}
}
\value{
expanded_dt
}
\description{
function takes a data table of reads
summarizes count of grouped reads
returns dataframe of top reads replicated to reflect their proportional weight to the top read
}
