% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_by_uservalue.R
\name{weight_by_uservalue}
\alias{weight_by_uservalue}
\title{Weight_by_uservalue
Takes a "collapsed" data.frame of reads summarized by count.
Replicates them according to a factor provided by the user.}
\usage{
weight_by_uservalue(dt, norm, locus_length)
}
\arguments{
\item{dt}{A dataframe of reads.}

\item{norm}{An integer chosen by the user which will be used to normalize the read counts.}

\item{locus_length}{An integer corresponding to the length of the region of interest}
}
\value{
expanded_dt A dataframe of reads which have been "uncollapsed" and replicated by the normalization factor.
}
\description{
Weight_by_uservalue
Takes a "collapsed" data.frame of reads summarized by count.
Replicates them according to a factor provided by the user.
}
