% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_proper_overhangs.R
\name{write_proper_overhangs}
\alias{write_proper_overhangs}
\title{write_proper_overhangs function
if write_fastas == TRUE, outputs read pairs with 3' 2nt overhang as a fasta file}
\usage{
write_proper_overhangs(dt1, dt2, wkdir, prefix, overlaps, suffix)
}
\arguments{
\item{dt1}{a data frame of reads}

\item{dt2}{a data frame of reads}

\item{wkdir}{a string specifying the directory to write the file to}

\item{prefix}{a string specifying the prefix of the file name}

\item{overlaps}{Specifies whether file types for plots are png or pdf. Default is pdf.}

\item{suffix}{a string specifying the suffix of the file name}
}
\value{
nothing
}
\description{
write_proper_overhangs function
if write_fastas == TRUE, outputs read pairs with 3' 2nt overhang as a fasta file
}
