% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_zscore.R
\name{calc_zscore}
\alias{calc_zscore}
\title{calc z_score
takes a data frame
returns z_score}
\usage{
calc_zscore(allCounts)
}
\arguments{
\item{allCounts}{a data frame}
}
\value{
z_score
}
\description{
calc z_score
takes a data frame
returns z_score
}
