% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_strand_hairpin.R
\name{dual_strand_hairpin}
\alias{dual_strand_hairpin}
\title{function to run the hairpin algorithm
processes reads from bam object according to strand
plots the arc plot and read dist}
\usage{
dual_strand_hairpin(
  chrom_name,
  reg_start,
  reg_stop,
  length,
  min_read_count,
  genome_file,
  bam_file,
  logfile,
  dir,
  plot_output,
  path_to_RNAfold,
  annotate_bed,
  weight_reads,
  bed_file
)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{length}{a whole number}

\item{min_read_count}{a whole number}

\item{genome_file}{a fasta file of chrom sequences}

\item{bam_file}{a BAM file}

\item{logfile}{a string}

\item{dir}{a string}

\item{plot_output}{a string, 'T' or 'F', default = 'T}

\item{path_to_RNAfold}{a string}

\item{annotate_bed}{a string, "T" or "F"}

\item{weight_reads}{a string, "T" or "F"}

\item{bed_file}{a string}
}
\value{
max_overhang
}
\description{
function to run the hairpin algorithm
processes reads from bam object according to strand
plots the arc plot and read dist
}
