% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_zscore.R
\name{get_max_zscore}
\alias{get_max_zscore}
\title{function to find max z_score
takes a table with one column of z_scores and one column of overlaps
returns max_zscore}
\usage{
get_max_zscore(z_score, Overlap)
}
\arguments{
\item{z_score}{a column of a data table containing z_scores}

\item{Overlap}{a column of a data table containing overlap values}
}
\value{
max_z
}
\description{
function to find max z_score
takes a table with one column of z_scores and one column of overlaps
returns max_zscore
}
