% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_dist.R
\name{get_read_dist}
\alias{get_read_dist}
\title{function to calculate the read distribution
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution}
\usage{
get_read_dist(chrom_m, chrom_p)
}
\arguments{
\item{chrom_m}{a minus strand object}

\item{chrom_p}{a plus strand object}
}
\value{
dizeDist
}
\description{
function to calculate the read distribution
takes forward and reverse data frames
summarizes read size by count
returns table of size distribution
}
