% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_n.R
\name{get_top_n}
\alias{get_top_n}
\title{function takes a data table of reads
summarizes count of grouped reads
returns an unweighted top n\% of reads}
\usage{
get_top_n(filter_dt, chrom_name, seq = NULL)
}
\arguments{
\item{chrom_name}{a string}

\item{seq}{a string, "T" or "F"}

\item{chrom}{a chrom object}
}
\value{
filter_dt
}
\description{
function takes a data table of reads
summarizes count of grouped reads
returns an unweighted top n\% of reads
}
