% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_n_weighted.R
\name{get_top_n_weighted}
\alias{get_top_n_weighted}
\title{function takes a data table of reads
summarizes count of grouped reads
returns dataframe of top reads replicated to reflect their weight}
\usage{
get_top_n_weighted(dt, chrom_name, seq = NULL)
}
\arguments{
\item{dt}{a data table of reads}

\item{chrom_name}{a string}

\item{seq}{a string, "T" or "F"}
}
\value{
filter_dt
}
\description{
function takes a data table of reads
summarizes count of grouped reads
returns dataframe of top reads replicated to reflect their weight
}
