% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChrPlus.R
\name{getChrPlus}
\alias{getChrPlus}
\title{create an object from read data from plus strand
takes bam obj, chrom name, reg_start, and reg_stop
extracts rname, pos, qwidth, and seq from bam obj}
\usage{
getChrPlus(bam_obj, chrom_name, reg_start, reg_stop)
}
\arguments{
\item{bam_obj}{a bam object created by openBamFile}

\item{chrom_name}{a name passed in by user/bed file}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}
}
\value{
chrom obj
}
\description{
create an object from read data from plus strand
takes bam obj, chrom name, reg_start, and reg_stop
extracts rname, pos, qwidth, and seq from bam obj
}
