% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_norm.R
\name{locus_norm}
\alias{locus_norm}
\title{normalize read counts by total number of reads mapped to locus}
\usage{
locus_norm(dt, locus_read_count)
}
\arguments{
\item{dt}{a dataframe of reads summarized by count}

\item{locus_read_count}{an integer}
}
\value{
final_df a dataframe
}
\description{
normalize read counts by total number of reads mapped to locus
}
