% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA_function.R
\name{miRNA_function}
\alias{miRNA_function}
\title{miRNA function
processes reads according to phased miRNA algorithm
plots output
takes dt and strand 
returns plots}
\usage{
miRNA_function(
  test_list,
  chrom_name,
  reg_start,
  reg_stop,
  bam_obj,
  bam_header,
  chr_name,
  chr_length,
  chromosome,
  strand,
  min_read_count,
  genome_file,
  input_file,
  dt
)
}
\arguments{
\item{test_list}{a bed file}

\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{bam_obj}{a bam object created by openBamFile}

\item{bam_header}{header information from bam file}

\item{chr_name}{chrom names extracted from bam file}

\item{chr_length}{length of chromosome of interest}

\item{chromosome}{index of chromosome of interest}

\item{strand}{a character passed in, "+" or "-"}

\item{min_read_count}{a whole number}

\item{genome_file}{a fasta file of chrom sequences}

\item{input_file}{a BAM file}

\item{dt}{a data table}
}
\value{
plots

max_overhang
}
\description{
miRNA function
processes reads according to phased miRNA algorithm
plots output
takes dt and strand 
returns plots
}
