% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_weight.R
\name{no_weight}
\alias{no_weight}
\title{function takes a data table of reads
summarizes count of grouped reads
returns an unweighted top n\% of reads}
\usage{
no_weight(filter_dt, chrom_name)
}
\arguments{
\item{chrom_name}{a string}

\item{chrom}{a chrom object}
}
\value{
filter_dt
}
\description{
function takes a data table of reads
summarizes count of grouped reads
returns an unweighted top n\% of reads
}
