% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenBamFile.R
\name{OpenBamFile}
\alias{OpenBamFile}
\title{function to open bam file and verify that it was successfully accessed
takes input file
creates bam obj}
\usage{
OpenBamFile(input_file)
}
\arguments{
\item{input_file}{a BAM file}
}
\value{
bam_obj
}
\description{
function to open bam file and verify that it was successfully accessed
takes input file
creates bam obj
}
