% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phasedz.R
\name{plot_phasedz}
\alias{plot_phasedz}
\title{phasing function
processes reads according to phased piRNA algorithm
plots output
takes data frame
returns plots}
\usage{
plot_phasedz(df1, strand = NULL)
}
\arguments{
\item{df1}{a data frame}

\item{strand}{a string, "+" or "-"}
}
\value{
plots
}
\description{
phasing function
processes reads according to phased piRNA algorithm
plots output
takes data frame
returns plots
}
