% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_si_heat.R
\name{plot_si_heat}
\alias{plot_si_heat}
\title{plot the si heatmaps
takes a matrix as input and a specified color palette
outputs plots}
\usage{
plot_si_heat(
  results,
  chrom_name,
  reg_start,
  reg_stop,
  wkdir,
  pal = c("RdYlBl", "yelOrRed", "MagYel", "Greens", "BlYel")
)
}
\arguments{
\item{results}{a data table}

\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{wkdir}{a string}

\item{pal}{a string}
}
\value{
plots
}
\description{
plot the si heatmaps
takes a matrix as input and a specified color palette
outputs plots
}
