% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sizes.R
\name{plot_sizes}
\alias{plot_sizes}
\title{plots the size distribution of reads}
\usage{
plot_sizes(dist)
}
\arguments{
\item{dist}{a data table}
}
\value{
plots
}
\description{
plots the size distribution of reads
}
