% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sizes_by_strand.R
\name{plot_sizes_by_strand}
\alias{plot_sizes_by_strand}
\title{plots the size distribution of reads separately by strand}
\usage{
plot_sizes_by_strand(dist)
}
\arguments{
\item{dist}{A data frame made by get_stranded_read_dist function.}
}
\value{
A plot
}
\description{
plots the size distribution of reads separately by strand
}
