% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_miRNA_function.R
\name{run_miRNA_function}
\alias{run_miRNA_function}
\title{run miRNA function
calls the miRNA function for both strands, creates miRNA logfile}
\usage{
run_miRNA_function(vars)
}
\arguments{
\item{vars}{a list of variables provided by user}
}
\value{
plots
}
\description{
run miRNA function
calls the miRNA function for both strands, creates miRNA logfile
}
