% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_miRNA_function.R
\name{run_miRNA_function}
\alias{run_miRNA_function}
\title{miRNA_function
processes reads according to phased miRNA algorithm
plots output
returns plots}
\usage{
run_miRNA_function(
  chrom_name,
  reg_start,
  reg_stop,
  chromosome,
  length,
  strand,
  min_read_count,
  genome_file,
  bam_file,
  logfile,
  wkdir,
  plot_output,
  path_to_RNAfold,
  weight_reads
)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{chromosome}{chrom names extracted from bam file}

\item{length}{length of chromosome of interest}

\item{strand}{a character passed in, "+" or "-"}

\item{min_read_count}{a whole number}

\item{genome_file}{a fasta file of chrom sequences}

\item{bam_file}{a BAM file}

\item{logfile}{a string}

\item{wkdir}{a string}

\item{plot_output}{a string, default = TRUE}

\item{path_to_RNAfold}{a string}

\item{weight_reads}{Determines whether read counts will be weighted and with which method. Valid options are "weight_by_prop", "locus_norm", a user-defined value, or "none". See MiSiPi documentation for descriptions of the weighting methods.}
}
\value{
plots
}
\description{
miRNA_function
processes reads according to phased miRNA algorithm
plots output
returns plots
}
