% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_phased_piRNA_function.R
\name{run_phased_piRNA}
\alias{run_phased_piRNA}
\title{run_phased_piRNA}
\usage{
run_phased_piRNA(bed_file, input_file, min_read_count)
}
\arguments{
\item{bed_file}{a tab-delineated file with three columns: chrom, start, stop}

\item{input_file}{a BAM file}

\item{min_read_count}{a whole number}
}
\value{
plots
}
\description{
run_phased_piRNA
}
