% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_piRNA_function.R
\name{run_piRNA_function}
\alias{run_piRNA_function}
\title{runs piRNA_function and creates dir and logfile for outputs}
\usage{
run_piRNA_function(vars)
}
\arguments{
\item{vars}{a list}
}
\value{
pi_res, plots
}
\description{
runs piRNA_function and creates dir and logfile for outputs
}
