% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_piRNA_function.R
\name{run_piRNA_function}
\alias{run_piRNA_function}
\title{Run the piRNA function
Finds overlapping ping-piRNAs, then performs single-strand phasing calculation
Outputs a plot, a table of values for a heatmap, and a table of z-scores.}
\usage{
run_piRNA_function(
  chrom_name,
  reg_start,
  reg_stop,
  length,
  bam_file,
  genome_file,
  logfile,
  wkdir,
  pal,
  plot_output,
  weight_reads,
  write_fastas,
  out_type
)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{length}{an integer}

\item{bam_file}{a string}

\item{genome_file}{a string}

\item{logfile}{a string}

\item{wkdir}{a string}

\item{pal}{a string}

\item{plot_output}{a bool, TRUE or FALSE, default = TRUE}

\item{weight_reads}{a string, Determines whether read counts will be weighted and with which method. Valid options are "weight_by_prop", "locus_norm", a user-defined value, or "none". See MiSiPi documentation for descriptions of the weighting methods.}

\item{write_fastas}{a bool, Determines whether piRNA pairs will be written to fasta. Expected values are TRUE or FALSE}

\item{out_type}{The type of file to write the plots to. Options are "png" or "pdf". Default is PDF.}
}
\value{
plots, heat results, and zdf
}
\description{
Run the piRNA function
Finds overlapping ping-piRNAs, then performs single-strand phasing calculation
Outputs a plot, a table of values for a heatmap, and a table of z-scores.
}
