% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_piRNA_function.R
\name{run_piRNA_function}
\alias{run_piRNA_function}
\title{run the piRNA function
processes forward and reverse reads according to piRNA algorithm
outputs plots, heat results, and zdf}
\usage{
run_piRNA_function(
  chrom_name,
  reg_start,
  reg_stop,
  length,
  bam_file,
  genome_file,
  logfile,
  wkdir,
  pal,
  plot_output
)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{length}{an integer}

\item{bam_file}{a string}

\item{genome_file}{a string}

\item{logfile}{a string}

\item{wkdir}{a string}

\item{pal}{a string}

\item{plot_output}{a string, 'T' or 'F', default = 'T}
}
\value{
plots, heat results, and zdf
}
\description{
run the piRNA function
processes forward and reverse reads according to piRNA algorithm
outputs plots, heat results, and zdf
}
