% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_piRNA_function.R
\name{run_piRNA_function}
\alias{run_piRNA_function}
\title{Run the piRNA function
Finds overlapping ping-piRNAs, then performs single-strand phasing calculation
Outputs a plot, a table of values for a heatmap, and a table of z-scores.}
\usage{
run_piRNA_function(
  chrom_name,
  reg_start,
  reg_stop,
  length,
  bam_file,
  genome_file,
  logfile,
  wkdir,
  pal,
  plot_output,
  weight_reads,
  write_fastas
)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{length}{an integer}

\item{bam_file}{a string}

\item{genome_file}{a string}

\item{logfile}{a string}

\item{wkdir}{a string}

\item{pal}{a string}

\item{plot_output}{a string, 'T' or 'F', default = 'T}

\item{weight_reads}{Determines whether read counts will be weighted and with which method. Valid options are "weight_by_prop", "locus_norm", a user-defined value, or "none". See MiSiPi documentation for descriptions of the weighting methods.}

\item{write_fastas}{Determines whether piRNA pairs will be written to fasta. Expected values are "T" or "F".}
}
\value{
plots, heat results, and zdf
}
\description{
Run the piRNA function
Finds overlapping ping-piRNAs, then performs single-strand phasing calculation
Outputs a plot, a table of values for a heatmap, and a table of z-scores.
}
