% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_siRNA_function.R
\name{run_siRNA_function}
\alias{run_siRNA_function}
\title{run siRNA function
process forward and reverse dt and plot si heatmap}
\usage{
run_siRNA_function(
  chrom_name,
  reg_start,
  reg_stop,
  length,
  min_read_count,
  genome_file,
  bam_file,
  logfile,
  wkdir,
  pal,
  plot_output,
  path_to_RNAfold,
  annotate_region,
  weight_reads,
  gtf_file,
  write_fastas
)
}
\arguments{
\item{chrom_name}{a string}

\item{reg_start}{a whole number}

\item{reg_stop}{a whole number}

\item{length}{an integer}

\item{min_read_count}{an integer}

\item{genome_file}{a string}

\item{bam_file}{a string}

\item{logfile}{a string}

\item{wkdir}{a string}

\item{pal}{a string}

\item{plot_output}{a string, 'T' or 'F'. Default is 'T'}

\item{path_to_RNAfold}{a string}

\item{annotate_region}{a string, "T" or "F"}

\item{weight_reads}{a string, "T" or "F"}

\item{gtf_file}{a string}

\item{write_fastas}{Determines whether siRNA pairs will be written to a fasta file. "T" or "F" expected.}
}
\value{
results
}
\description{
run siRNA function
process forward and reverse dt and plot si heatmap
}
