% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_siRNA_function.R
\name{run_siRNA_function}
\alias{run_siRNA_function}
\title{run_siRNA_function}
\usage{
run_siRNA_function(vars)
}
\arguments{
\item{vars}{a list}
}
\value{
si_res
}
\description{
run_siRNA_function
}
