% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siRNA_function.R
\name{siRNA_function}
\alias{siRNA_function}
\title{run siRNA function
calls the siRNA function, creates siRNA logfile}
\usage{
siRNA_function(vars)
}
\arguments{
\item{vars}{a list of variables provided by user}
}
\value{
si_res
}
\description{
run siRNA function
calls the siRNA function, creates siRNA logfile
}
